+++ test/apm-server-version.test.js
@@ -42,6 +42,8 @@ test('APM server version fetch works for "6.6.0"', function (t) {
       'client._apmServerVersion is undefined immediately after client creation')
     t.equal(client.supportsKeepingUnsampledTransaction(), true,
       'client.supportsKeepingUnsampledTransaction() defaults to true before fetch')
+    t.equal(client.supportsActivationMethodField(), false,
+      'client.supportsActivationMethodField() defaults to false before fetch')
 
     // Currently there isn't a mechanism to wait for the fetch request, so for
     // now just wait a bit.
@@ -50,6 +52,8 @@ test('APM server version fetch works for "6.6.0"', function (t) {
       t.equal(client._apmServerVersion.toString(), '6.6.0')
       t.equal(client.supportsKeepingUnsampledTransaction(), true,
         'client.supportsKeepingUnsampledTransaction() is true after fetch')
+      t.equal(client.supportsActivationMethodField(), false,
+        'client.supportsActivationMethodField() is false after fetch')
 
       server.close()
       client.destroy()
@@ -116,6 +120,90 @@ test('APM server version fetch works for "8.0.0"', function (t) {
       t.equal(client._apmServerVersion.toString(), '8.0.0')
       t.equal(client.supportsKeepingUnsampledTransaction(), false,
         'client.supportsKeepingUnsampledTransaction() is false after fetch')
+
+      server.close()
+      client.destroy()
+      t.end()
+    }, 200)
+  })
+})
+
+/**
+ * APM server 8.7.0 included a bug where APM agents sending `activation_method`
+ * was harmful. This test ensures we don't send that field to v8.7.0.
+ *
+ * See https://github.com/elastic/apm/pull/780
+ */
+test('APM server version fetch works for "8.7.0"', function (t) {
+  const server = APMServer(function (req, res) {
+    res.writeHead(200)
+    const verInfo = {
+      build_date: '2023-03-30T22:17:50Z',
+      build_sha: 'a183f675ecd03fca4a897cbe85fda3511bc3ca43',
+      version: '8.7.0'
+    }
+    res.end(JSON.stringify(verInfo, null, 2))
+  }).client({
+    agentActivationMethod: 'env-attach'
+  }, function (client) {
+    t.strictEqual(client._apmServerVersion, undefined,
+      'client._apmServerVersion is undefined immediately after client creation')
+    t.equal(client._conf.agentActivationMethod, 'env-attach', '_conf.agentActivationMethod')
+    t.equal(client.supportsActivationMethodField(), false,
+      'client.supportsActivationMethodField() defaults to false before fetch')
+    t.equal(JSON.parse(client._encodedMetadata).metadata.service.agent.activation_method, undefined,
+      'metadata does not include "activation_method" before fetch')
+
+    // Currently there isn't a mechanism to wait for the fetch request, so for
+    // now just wait a bit.
+    setTimeout(() => {
+      t.ok(client._apmServerVersion, 'client._apmServerVersion is set')
+      t.equal(client._apmServerVersion.toString(), '8.7.0')
+      t.equal(client.supportsActivationMethodField(), false,
+        'client.supportsActivationMethodField() is false after fetch')
+      t.equal(JSON.parse(client._encodedMetadata).metadata.service.agent.activation_method, undefined,
+        'metadata does not include "activation_method" after fetch')
+
+      server.close()
+      client.destroy()
+      t.end()
+    }, 200)
+  })
+})
+
+/**
+ * Starting with APM server 8.7.1, `activation_method` should be sent.
+ * See https://github.com/elastic/apm/pull/780
+ */
+test('APM server version fetch works for "8.7.1"', function (t) {
+  const server = APMServer(function (req, res) {
+    res.writeHead(200)
+    const verInfo = {
+      build_date: '2023-03-30T22:17:50Z',
+      build_sha: 'a183f675ecd03fca4a897cbe85fda3511bc3ca43',
+      version: '8.7.1'
+    }
+    res.end(JSON.stringify(verInfo, null, 2))
+  }).client({
+    agentActivationMethod: 'env-attach'
+  }, function (client) {
+    t.strictEqual(client._apmServerVersion, undefined,
+      'client._apmServerVersion is undefined immediately after client creation')
+    t.equal(client._conf.agentActivationMethod, 'env-attach', '_conf.agentActivationMethod')
+    t.equal(client.supportsActivationMethodField(), false,
+      'client.supportsActivationMethodField() defaults to false before fetch')
+    t.equal(JSON.parse(client._encodedMetadata).metadata.service.agent.activation_method, undefined,
+      'metadata does not include "activation_method" before fetch')
+
+    // Currently there isn't a mechanism to wait for the fetch request, so for
+    // now just wait a bit.
+    setTimeout(() => {
+      t.ok(client._apmServerVersion, 'client._apmServerVersion is set')
+      t.equal(client._apmServerVersion.toString(), '8.7.1')
+      t.equal(client.supportsActivationMethodField(), true,
+        'client.supportsActivationMethodField() is true after fetch')
+      t.equal(JSON.parse(client._encodedMetadata).metadata.service.agent.activation_method, 'env-attach',
+        'metadata includes "activation_method" after fetch')
 
       server.close()
       client.destroy()